SniffyStackoverflowApplicationTests moved to sniffy-core
