--- src/main/resources/application.properties
@@ -1,1 +1,0 @@
-logging.level.org.springframework.jdbc.core.JdbcTemplate = debug
